+++ integration/helpers.go
@@ -155,6 +155,53 @@ func AssertEqualAltError(t testing.TB, expected mongo.CommandError, altMessage s
 	return assert.Equal(t, expected, a)
 }
 
+// AssertEqualWriteError compares expected mongo.WriteError Message and Code with actual error.
+func AssertEqualWriteError(t *testing.T, expected *mongo.WriteError, actual error) bool {
+	t.Helper()
+
+	writeException, ok := actual.(mongo.WriteException)
+	if !ok {
+		return assert.Equal(t, expected, actual)
+	}
+
+	if len(writeException.WriteErrors) != 1 {
+		return assert.Equal(t, expected, actual)
+	}
+
+	actualWriteErr := writeException.WriteErrors[0]
+
+	return assert.Equal(t, expected.Message, actualWriteErr.Message) &&
+		assert.Equal(t, expected.Code, actualWriteErr.Code)
+}
+
+// AssertEqualAltWriteError compares expected mongo.WriteError Message and Code with actual error.
+func AssertEqualAltWriteError(t *testing.T, expected *mongo.WriteError, alt string, actual error) bool {
+	t.Helper()
+
+	var actualWE mongo.WriteError
+	actualEx, ok := actual.(mongo.WriteException)
+	if ok {
+		if len(actualEx.WriteErrors) != 1 {
+			return assert.Equal(t, expected, actual)
+		}
+		actualWE = actualEx.WriteErrors[0]
+	} else {
+		actualWE, ok = actual.(mongo.WriteError)
+		if !ok {
+			return assert.Equal(t, expected, actual)
+		}
+	}
+	assert.Equal(t, int(expected.Code), int(actualWE.Code), "codes are not equal", expected.Code, actualWE.Code)
+	if assert.ObjectsAreEqual(expected.Message, actualWE.Message) {
+		return true
+	}
+	if alt != "" {
+		expected.Message = alt
+	}
+	return assert.Equal(t, expected.Message, actualWE.Message) &&
+		assert.Equal(t, expected.Code, actualWE.Code)
+}
+
 // CollectIDs returns all _id values from given documents.
 //
 // The order is preserved.
@@ -183,70 +230,4 @@ func CollectKeys(t testing.TB, doc bson.D) []string {
 	}
 
 	return res
-}
-
-// AssertEqualWriteError compares expected mongo.WriteError Message and Code with actual error.
-func AssertEqualWriteError(t *testing.T, expected *mongo.WriteError, actual error) bool {
-	t.Helper()
-
-	writeException, ok := actual.(mongo.WriteException)
-	if !ok {
-		return assert.Equal(t, expected, actual)
-	}
-
-	if len(writeException.WriteErrors) != 1 {
-		return assert.Equal(t, expected, actual)
-	}
-
-	actualWriteErr := writeException.WriteErrors[0]
-
-	return assert.Equal(t, expected.Message, actualWriteErr.Message) &&
-		assert.Equal(t, expected.Code, actualWriteErr.Code)
-}
-
-// IsUnimplemented returns true if the error code is unimplemented.
-func IsUnimplemented(t *testing.T, actual error) bool {
-	actualCE, ok := actual.(mongo.CommandError)
-	if ok {
-		if actualCE.Code == 238 {
-			return true
-		}
-	}
-	return false
-}
-
-// AssertEqualAltWriteError compares expected mongo.WriteError Message and Code with actual error.
-func AssertEqualAltWriteError(t *testing.T, expected *mongo.WriteError, alt string, actual error) bool {
-	t.Helper()
-
-	var actualWE mongo.WriteError
-	actualEx, ok := actual.(mongo.WriteException)
-	if ok {
-		if len(actualEx.WriteErrors) != 1 {
-			return assert.Equal(t, expected, actual)
-		}
-		actualWE = actualEx.WriteErrors[0]
-	} else {
-		actualWE, ok = actual.(mongo.WriteError)
-		if !ok {
-			t.Log("compare", expected, actual)
-			return assert.Equal(t, expected, actual)
-		}
-	}
-
-	if int(expected.Code) != int(actualWE.Code) {
-		t.Log("codes are not equal", expected.Code, actualWE.Code)
-		return false
-	}
-	if expected.Message == actualWE.Message {
-		return true
-	}
-	t.Log(expected.Message)
-	t.Log(actualWE.Message)
-	t.Log(alt)
-	if alt != "" {
-		expected.Message = alt
-		return expected.Message == actualWE.Message
-	}
-	return false
 }
