+++ integration/update_test.go
@@ -282,3 +282,473 @@ func TestUpdateIncOperator(t *testing.T) {
 		})
 	}
 }
+
+func TestUpdateSet(t *testing.T) {
+	t.Parallel()
+
+	for name, tc := range map[string]struct {
+		id     string
+		update bson.D
+		result bson.D
+		err    *mongo.WriteError
+		stat   *mongo.UpdateResult
+		alt    string
+	}{
+		"Many": {
+			id:     "string",
+			update: bson.D{{"$set", bson.D{{"foo", int32(1)}, {"bar", bson.A{}}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "string"}, {"value", "foo"}, {"bar", bson.A{}}, {"foo", int32(1)}},
+		},
+		"NilOperand": {
+			id:     "string",
+			update: bson.D{{"$set", nil}},
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type null instead. " +
+					"For example: {$mod: {<field>: ...}} not {$set: null}",
+			},
+		},
+		"String": {
+			id:     "string",
+			update: bson.D{{"$set", "string"}},
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type string instead. " +
+					"For example: {$mod: {<field>: ...}} not {$set: \"string\"}",
+			},
+			alt: "Modifiers operate on fields but we found type string instead. " +
+				"For example: {$mod: {<field>: ...}} not {$set: string}",
+		},
+		"Array": {
+			id:     "string",
+			update: bson.D{{"$set", bson.A{}}},
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type array instead. " +
+					"For example: {$mod: {<field>: ...}} not {$set: []}",
+			},
+			alt: "Modifiers operate on fields but we found type array instead. " +
+				"For example: {$mod: {<field>: ...}} not {$set: array}",
+		},
+		"EmptyDoc": {
+			id:     "string",
+			update: bson.D{{"$set", bson.D{}}},
+			result: bson.D{{"_id", "string"}, {"value", "foo"}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 0,
+				UpsertedCount: 0,
+			},
+		},
+		"OkSetString": {
+			id:     "string",
+			update: bson.D{{"$set", bson.D{{"value", "ok value"}}}},
+			result: bson.D{{"_id", "string"}, {"value", "ok value"}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+		},
+		"ArrayNil": {
+			id:     "string",
+			update: bson.D{{"$set", bson.D{{"value", bson.A{nil}}}}},
+			result: bson.D{{"_id", "string"}, {"value", bson.A{nil}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+		},
+		"FieldNotExist": {
+			id:     "string",
+			update: bson.D{{"$set", bson.D{{"foo", int32(1)}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "string"}, {"value", "foo"}, {"foo", int32(1)}},
+		},
+		"Double": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", float64(1)}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", float64(1)}},
+		},
+		"NaN": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", math.NaN()}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", math.NaN()}},
+		},
+		"EmptyArray": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", bson.A{}}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", bson.A{}}},
+		},
+		"Null": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", nil}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", nil}},
+		},
+		"Int32": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", int32(1)}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", int32(1)}},
+		},
+		"Inf": {
+			id:     "double",
+			update: bson.D{{"$set", bson.D{{"value", math.Inf(+1)}}}},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			result: bson.D{{"_id", "double"}, {"value", math.Inf(+1)}},
+		},
+	} {
+		name, tc := name, tc
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			ctx, collection := setup(t)
+			_, err := collection.InsertMany(ctx, []any{
+				bson.D{{"_id", "string"}, {"value", "foo"}},
+				bson.D{{"_id", "double"}, {"value", float64(0.0)}},
+			})
+			require.NoError(t, err)
+
+			res, err := collection.UpdateOne(ctx, bson.D{{"_id", tc.id}}, tc.update)
+			if tc.err != nil {
+				require.Nil(t, tc.result)
+				if !AssertEqualAltWriteError(t, tc.err, tc.alt, err) {
+					t.FailNow()
+				}
+				return
+			}
+
+			require.NoError(t, err)
+			require.Equal(t, tc.stat, res)
+
+			var actual bson.D
+			err = collection.FindOne(ctx, bson.D{{"_id", tc.id}}).Decode(&actual)
+			require.NoError(t, err)
+			if !AssertEqualDocuments(t, tc.result, actual) {
+				t.FailNow()
+			}
+		})
+	}
+}
+
+func TestSetOnInsert(t *testing.T) {
+	t.Parallel()
+
+	notModified := &mongo.UpdateResult{
+		MatchedCount:  0,
+		ModifiedCount: 0,
+		UpsertedCount: 1,
+	}
+
+	for name, tc := range map[string]struct {
+		filter      bson.D
+		setOnInsert any
+		stat        *mongo.UpdateResult
+		res         bson.D
+		err         *mongo.WriteError
+		alt         string
+	}{
+		"doc": {
+			filter:      bson.D{{"_id", "doc"}},
+			setOnInsert: bson.D{{"value", bson.D{}}},
+			res:         bson.D{{"_id", "doc"}, {"value", bson.D{}}},
+		},
+		"array": {
+			filter:      bson.D{{"_id", "array"}},
+			setOnInsert: bson.D{{"value", bson.A{}}},
+			res:         bson.D{{"_id", "array"}, {"value", bson.A{}}},
+		},
+		"double": {
+			filter:      bson.D{{"_id", "double"}},
+			setOnInsert: bson.D{{"value", 43.13}},
+			res:         bson.D{{"_id", "double"}, {"value", 43.13}},
+		},
+		"NaN": {
+			filter:      bson.D{{"_id", "double-nan"}},
+			setOnInsert: bson.D{{"value", math.NaN()}},
+			res:         bson.D{{"_id", "double-nan"}, {"value", math.NaN()}},
+		},
+		"string": {
+			filter:      bson.D{{"_id", "string"}},
+			setOnInsert: bson.D{{"value", "abcd"}},
+			res:         bson.D{{"_id", "string"}, {"value", "abcd"}},
+		},
+		"nil": {
+			filter:      bson.D{{"_id", "nil"}},
+			setOnInsert: bson.D{{"value", nil}},
+			res:         bson.D{{"_id", "nil"}, {"value", nil}},
+		},
+		"empty-doc": {
+			filter:      bson.D{{"_id", "doc"}},
+			setOnInsert: bson.D{},
+			res:         bson.D{{"_id", "doc"}},
+		},
+		"empty-array": {
+			filter:      bson.D{{"_id", "array"}},
+			setOnInsert: bson.A{},
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type array instead. " +
+					"For example: {$mod: {<field>: ...}} not {$setOnInsert: []}",
+			},
+			alt: "Modifiers operate on fields but we found type array instead. " +
+				"For example: {$mod: {<field>: ...}} not {$setOnInsert: array}",
+		},
+		"double-double": {
+			filter:      bson.D{{"_id", "double"}},
+			setOnInsert: 43.13,
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type double instead. " +
+					"For example: {$mod: {<field>: ...}} not {$setOnInsert: 43.13}",
+			},
+			alt: "Modifiers operate on fields but we found type double instead. " +
+				"For example: {$mod: {<field>: ...}} not {$setOnInsert: double}",
+		},
+		"err-NaN": {
+			filter:      bson.D{{"_id", "double-nan"}},
+			setOnInsert: math.NaN(),
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type double instead. " +
+					"For example: {$mod: {<field>: ...}} not {$setOnInsert: nan.0}",
+			},
+			alt: "Modifiers operate on fields but we found type double instead. " +
+				"For example: {$mod: {<field>: ...}} not {$setOnInsert: double}",
+		},
+		"err-string": {
+			filter:      bson.D{{"_id", "string"}},
+			setOnInsert: "any string",
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type string instead. " +
+					"For example: {$mod: {<field>: ...}} not {$setOnInsert: \"any string\"}",
+			},
+			alt: "Modifiers operate on fields but we found type string instead. " +
+				"For example: {$mod: {<field>: ...}} not {$setOnInsert: string}",
+		},
+		"err-nil": {
+			filter:      bson.D{{"_id", "nil"}},
+			setOnInsert: nil,
+			err: &mongo.WriteError{
+				Code: 9,
+				Message: "Modifiers operate on fields but we found type null instead. " +
+					"For example: {$mod: {<field>: ...}} not {$setOnInsert: null}",
+			},
+			alt: "Modifiers operate on fields but we found type null instead. " +
+				"For example: {$mod: {<field>: ...}} not {$setOnInsert: null}",
+		},
+	} {
+		name, tc := name, tc
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			var err error
+			ctx, collection := setup(t)
+
+			opts := options.Update().SetUpsert(true)
+			var res *mongo.UpdateResult
+			res, err = collection.UpdateOne(ctx, tc.filter, bson.D{{"$setOnInsert", tc.setOnInsert}}, opts)
+			if tc.err != nil {
+				if !AssertEqualAltWriteError(t, tc.err, tc.alt, err) {
+					t.Logf("%[1]T %[1]v", err)
+					t.FailNow()
+				}
+				return
+			}
+
+			require.NoError(t, err)
+			id := res.UpsertedID
+			assert.NotEmpty(t, id)
+			res.UpsertedID = nil
+			expectedRes := notModified
+			if tc.stat != nil {
+				expectedRes = tc.stat
+			}
+			assert.Equal(t, expectedRes, res)
+
+			var actual bson.D
+			err = collection.FindOne(ctx, tc.filter).Decode(&actual)
+			require.NoError(t, err)
+			if !AssertEqualDocuments(t, tc.res, actual) {
+				t.FailNow()
+			}
+		})
+	}
+}
+
+func TestSetOnInsertMore(t *testing.T) {
+	t.Parallel()
+
+	notModified := &mongo.UpdateResult{
+		MatchedCount:  0,
+		ModifiedCount: 0,
+		UpsertedCount: 1,
+	}
+
+	for name, tc := range map[string]struct {
+		filter         bson.D
+		query          bson.D
+		stat           *mongo.UpdateResult
+		res            bson.D
+		err            *mongo.WriteError
+		notImplemented bool
+		alt            string
+	}{
+		"tandem-set-setoninsert": {
+			filter: bson.D{{"_id", "test"}},
+			query: bson.D{
+				{"$set", bson.D{{"foo", int32(12)}}},
+				{"$setOnInsert", bson.D{{"value", math.NaN()}}},
+			},
+			res: bson.D{{"_id", "test"}, {"foo", int32(12)}, {"value", math.NaN()}},
+		},
+		"trio": {
+			filter: bson.D{{"_id", "test"}},
+			query: bson.D{
+				{"$set", bson.D{{"foo", int32(12)}}},
+				{"$inc", bson.D{{"foo", int32(1)}}},
+				{"$setOnInsert", bson.D{{"value", math.NaN()}}},
+			},
+			err: &mongo.WriteError{
+				Code:    40,
+				Message: "Updating the path 'foo' would create a conflict at 'foo'",
+			},
+		},
+		"unknown-operator": {
+			filter: bson.D{{"_id", "test"}},
+			query:  bson.D{{"$foo", bson.D{{"foo", int32(1)}}}},
+			err: &mongo.WriteError{
+				Code:    9,
+				Message: "Unknown modifier: $foo. Expected a valid update modifier or pipeline-style update specified as an array",
+			},
+		},
+		"set-dot-notation": {
+			filter: bson.D{{"_id", "string"}},
+			query:  bson.D{{"$set", bson.D{{"foo.bar.baz", int32(1)}}}},
+			res: bson.D{
+				{"_id", "string"},
+				{"value", "foo"},
+				{"foo", bson.D{{"bar", bson.D{{"baz", int32(1)}}}}},
+			},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  1,
+				ModifiedCount: 1,
+				UpsertedCount: 0,
+			},
+			notImplemented: true,
+		},
+		"setoninsert-dot-notation": {
+			filter:         bson.D{{"_id", "test"}},
+			query:          bson.D{{"$setOnInsert", bson.D{{"foo.bar.baz", int32(1)}}}},
+			res:            bson.D{{"_id", "test"}, {"foo", bson.D{{"bar", bson.D{{"baz", int32(1)}}}}}},
+			notImplemented: true,
+		},
+		"inc-dot-notation": {
+			filter: bson.D{{"_id", "int32"}},
+			query:  bson.D{{"$inc", bson.D{{"value.foo.bar.baz", int32(1)}}}},
+			res: bson.D{
+				{"_id", "int32"},
+				{"value", bson.D{{"foo", bson.D{{"bar", bson.D{{"baz", int32(1)}}}}}}},
+			},
+			stat: &mongo.UpdateResult{
+				MatchedCount:  0,
+				ModifiedCount: 0,
+				UpsertedCount: 1,
+			},
+			notImplemented: true,
+		},
+	} {
+		name, tc := name, tc
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			ctx, collection := setup(t)
+			var err error
+			_, err = collection.InsertMany(ctx, []any{
+				bson.D{{"_id", "string"}, {"value", "foo"}},
+			})
+			require.NoError(t, err)
+
+			opts := options.Update().SetUpsert(true)
+			var res *mongo.UpdateResult
+			res, err = collection.UpdateOne(ctx, tc.filter, tc.query, opts)
+
+			if tc.notImplemented {
+				if dbByPort() == "ferretdb" &&
+					isUnimplemented(t, err) {
+					return
+				}
+			}
+			if tc.err != nil {
+				if !AssertEqualAltWriteError(t, tc.err, tc.alt, err) {
+					t.Logf("%[1]T %[1]v", err)
+					t.FailNow()
+				}
+				return
+			}
+
+			require.NoError(t, err)
+			res.UpsertedID = nil
+
+			expectedRes := notModified
+			if tc.stat != nil {
+				expectedRes = tc.stat
+			}
+			assert.Equal(t, expectedRes, res)
+
+			var actual bson.D
+			err = collection.FindOne(ctx, tc.filter).Decode(&actual)
+			require.NoError(t, err)
+			if !AssertEqualDocuments(t, tc.res, actual) {
+				t.FailNow()
+			}
+		})
+	}
+}
+
+// isUnimplemented returns true if the error code is unimplemented.
+func isUnimplemented(t *testing.T, actual error) bool {
+	actualCE, ok := actual.(mongo.CommandError)
+	if ok {
+		if actualCE.Code == 238 {
+			return true
+		}
+	}
+	return false
+}
