+++ internal/handlers/common/update.go
@@ -23,6 +23,108 @@ import (
 	"github.com/FerretDB/FerretDB/internal/util/must"
 )
 
+// UpdateDocument updates the given document with a series of update operators.
+// Returns true if document was changed.
+func UpdateDocument(doc, update *types.Document) (bool, error) {
+	var changed bool
+	for _, updateOp := range update.Keys() {
+		updateV := must.NotFail(update.Get(updateOp))
+
+		switch updateOp {
+		case "$set",
+			"$setOnInsert":
+
+			switch setDoc := updateV.(type) {
+			case *types.Document:
+				if setDoc.Len() == 0 {
+					return false, nil
+				}
+				sort.Strings(setDoc.Keys())
+				for _, setKey := range setDoc.Keys() {
+					setValue := must.NotFail(setDoc.Get(setKey))
+					if err := doc.Set(setKey, setValue); err != nil {
+						return false, err
+					}
+				}
+				changed = true
+			default:
+				msgFmt := fmt.Sprintf(`Modifiers operate on fields but we found type %[1]s instead. `+
+					`For example: {$mod: {<field>: ...}} not {$set: %[1]s}`,
+					AliasFromType(updateV),
+				)
+				return false, NewWriteErrorMsg(ErrFailedToParse, msgFmt)
+			}
+
+		case "$inc":
+			incDoc, ok := updateV.(*types.Document)
+			if !ok {
+				return false, NewWriteErrorMsg(
+					ErrFailedToParse,
+					fmt.Sprintf(
+						`Modifiers operate on fields but we found type string instead. `+
+							`For example: {$mod: {<field>: ...}} not {%s: %#v}`,
+						updateOp,
+						updateV,
+					),
+				)
+			}
+
+			for _, incKey := range incDoc.Keys() {
+				if strings.ContainsRune(incKey, '.') {
+					return false, NewErrorMsg(ErrNotImplemented, "dot notation not supported yet")
+				}
+
+				incValue := must.NotFail(incDoc.Get(incKey))
+
+				if !doc.Has(incKey) {
+					must.NoError(doc.Set(incKey, incValue))
+					changed = true
+					continue
+				}
+
+				docValue := must.NotFail(doc.Get(incKey))
+
+				incremented, err := addNumbers(incValue, docValue)
+				if err == nil {
+					must.NoError(doc.Set(incKey, incremented))
+					changed = true
+					continue
+				}
+
+				switch err {
+				case errUnexpectedLeftOpType:
+					return false, NewWriteErrorMsg(
+						ErrTypeMismatch,
+						fmt.Sprintf(
+							`Cannot increment with non-numeric argument: {%s: %#v}`,
+							incKey,
+							incValue,
+						),
+					)
+				case errUnexpectedRightOpType:
+					return false, NewWriteErrorMsg(
+						ErrTypeMismatch,
+						fmt.Sprintf(
+							`Cannot apply $inc to a value of non-numeric type. `+
+								`{_id: "%s"} has the field '%s' of non-numeric type %s`,
+							must.NotFail(doc.Get("_id")),
+							incKey,
+							AliasFromType(docValue),
+						),
+					)
+				default:
+					return false, err
+				}
+			}
+
+		default:
+			return false, NewError(ErrNotImplemented, fmt.Errorf("UpdateDocument: unhandled operation %q", updateOp))
+		}
+	}
+
+	return changed, nil
+}
+
 // ValidateUpdateQuery validates update statement.
 func ValidateUpdateQuery(update *types.Document) (err error) {
 	set, err := getDocument("$set", update)
@@ -93,111 +195,4 @@ func getDocument(op string, update *types.Document) (*types.Document, error) {
 		)
 		return nil, NewWriteErrorMsg(ErrFailedToParse, msgFmt)
 	}
-}
-
-// UpdateDocument updates the given document with a series of update operators.
-// Returns true if document was changed.
-func UpdateDocument(doc, update *types.Document) (bool, error) {
-	var changed bool
-	for _, updateOp := range update.Keys() {
-		updateV := must.NotFail(update.Get(updateOp))
-
-		switch updateOp {
-		case "$set",
-			"$setOnInsert":
-			switch setDoc := updateV.(type) {
-			case *types.Document:
-				if setDoc.Len() == 0 {
-					return changed, nil
-				}
-				var err error
-				sort.Strings(setDoc.Keys())
-				for _, setKey := range setDoc.Keys() {
-					setValue := must.NotFail(setDoc.Get(setKey))
-					if err = doc.Set(setKey, setValue); err != nil {
-						return false, err
-					}
-				}
-				changed = true
-			default:
-				msgFmt := fmt.Sprintf(`Modifiers operate on fields but we found type %[1]s instead. `+
-					`For example: {$mod: {<field>: ...}} not {%s: %[1]s}`,
-					AliasFromType(updateV),
-					updateOp,
-				)
-				return false, NewWriteErrorMsg(ErrFailedToParse, msgFmt)
-			}
-
-		case "$inc":
-			incDoc, ok := updateV.(*types.Document)
-			if !ok {
-				return false, NewWriteErrorMsg(
-					ErrFailedToParse,
-					fmt.Sprintf(
-						`Modifiers operate on fields but we found type string instead. `+
-							`For example: {$mod: {<field>: ...}} not {%s: %#v}`,
-						updateOp,
-						updateV,
-					),
-				)
-			}
-
-			for _, incKey := range incDoc.Keys() {
-				if strings.ContainsRune(incKey, '.') {
-					return changed, NewErrorMsg(ErrNotImplemented, "dot notation not supported yet")
-				}
-
-				incValue := must.NotFail(incDoc.Get(incKey))
-
-				if !doc.Has(incKey) {
-					must.NoError(doc.Set(incKey, incValue))
-					changed = true
-					continue
-				}
-
-				docValue := must.NotFail(doc.Get(incKey))
-
-				incremented, err := addNumbers(incValue, docValue)
-				if err == nil {
-					must.NoError(doc.Set(incKey, incremented))
-					changed = true
-					continue
-				}
-
-				switch err {
-				case errUnexpectedLeftOpType:
-					return false, NewWriteErrorMsg(
-						ErrTypeMismatch,
-						fmt.Sprintf(
-							`Cannot increment with non-numeric argument: {%s: %#v}`,
-							incKey,
-							incValue,
-						),
-					)
-				case errUnexpectedRightOpType:
-					return false, NewWriteErrorMsg(
-						ErrTypeMismatch,
-						fmt.Sprintf(
-							`Cannot apply $inc to a value of non-numeric type. `+
-								`{_id: "%s"} has the field '%s' of non-numeric type %s`,
-							must.NotFail(doc.Get("_id")),
-							incKey,
-							AliasFromType(docValue),
-						),
-					)
-				default:
-					return false, err
-				}
-			}
-
-		default:
-			return false, NewWriteErrorMsg(
-				ErrFailedToParse,
-				fmt.Sprintf("Unknown modifier: %s. Expected a valid update modifier or pipeline-style "+
-					"update specified as an array", updateOp,
-				))
-		}
-	}
-
-	return changed, nil
 }
