+++ internal/handlers/pg/msg_update.go
@@ -166,8 +166,7 @@ func (h *Handler) MsgUpdate(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, 
 		matched += int32(len(resDocs))
 
 		for _, doc := range resDocs {
-			var changed bool
-			changed, err = common.UpdateDocument(doc, u)
+			changed, err := common.UpdateDocument(doc, u)
 			if err != nil {
 				return nil, err
 			}
